adk/requirements.txt
# Google ADK and AI
google-genai==0.3.0
google-adk==0.3.0

# PDF generation
fpdf2==2.7.9

# Data validation
pydantic==2.9.2

# Development tools
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1